+++ tests/testthat/test-keyATMCovPG.R
@@ -19,7 +19,7 @@ cov <- keyATM(docs = keyATM_docs,
               keywords = bills_keywords,
               model = "covariates",
               model_settings = list(covariates_data = bills_cov, standardize = "all", 
-                                  covariates_formula = ~.
+                                  covariates_formula = ~., covariates_model = "PG"
                                   ),
               options = list(seed = 250, store_theta = TRUE, iterations = 20,
                              store_pi = 1, thinning = 5, verbose = FALSE),
@@ -32,10 +32,10 @@ test_that("keyATM covariate", {
   expect_error(covariates_info(base))
 
   skip_on_os("linux") ; skip_on_cran()
-  expect_equal(cov$model_fit$Perplexity[3], 1874.663, tolerance = 0.001)
+  expect_equal(cov$model_fit$Perplexity[3], 1984.968, tolerance = 0.001)
   expect_equal(top_words(cov)[1, 1], "education [\U2713]")
   expect_equal(top_words(cov)[3, 3], "care")
-  expect_equal(cov$pi$Proportion[2], 4.836863, tolerance = 0.00001)
+  expect_equal(cov$pi$Proportion[2], 5.0288, tolerance = 0.00001)
 })
 
 
@@ -44,7 +44,20 @@ test_that("keyATM Heterogeneity Doc-Topic", {
   strata_topic <- by_strata_DocTopic(cov, by_var = "RepParty", labels = c("Dem", "Rep"), parallel = FALSE, posterior_mean = FALSE)
 
   skip_on_os("linux") ; skip_on_cran()
-  expect_equal(summary(strata_topic, method = "eti")[[2]]$Lower[2], 0.13578, tolerance = 0.00001)
+  expect_equal(summary(strata_topic, method = "eti")[[2]]$Lower[2], 0.07480147, tolerance = 0.00001)
+
+  p <- plot(strata_topic, show_topic = c(1,2,3,4), by = "covariate", method = "eti")
+  expect_s3_class(p, "keyATM_fig")
+
+  expect_message(suppressWarnings(save_fig(p, paste0(tempdir(), "/test.pdf"))), "Saving 7 x 7 in image")
+})
+
+
+test_that("keyATM Heterogeneity Doc-Topic, use posterior_mean", {
+  strata_topic <- by_strata_DocTopic(cov, by_var = "RepParty", labels = c("Dem", "Rep"), parallel = FALSE, posterior_mean = TRUE)
+
+  skip_on_os("linux") ; skip_on_cran()
+  expect_equal(summary(strata_topic, method = "eti")[[2]]$Lower[2], 0.1382071, tolerance = 0.00001)
 
   p <- plot(strata_topic, show_topic = c(1,2,3,4), by = "covariate", method = "eti")
   expect_s3_class(p, "keyATM_fig")
@@ -72,7 +85,7 @@ cov <- suppressWarnings(keyATM(docs = keyATM_docs,
               keywords = bills_keywords,
               model = "covariates",
               model_settings = list(covariates_data = bills_cov, standardize = "none", 
-                                  covariates_formula = NULL
+                                  covariates_formula = NULL, covariates_model = "PG"
                                   ),
               options = list(seed = 250, store_theta = TRUE, iterations = 5,
                              store_pi = 1, thinning = 5, verbose = FALSE)
@@ -84,12 +97,12 @@ test_that("Covariates settings: Standardize - none, no formula", {
   
   skip_on_os("linux") ; skip_on_cran()
   expect_error(predict(cov, bills_cov_modified))
-  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov, transform = TRUE))[3, 3]), 0.1734721, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov, transform = TRUE))[3, 3]), 0.1329791, tolerance = 0.000001)
 
   bills_cov_copy <- bills_cov
   bills_cov_copy[, 1] <- 1
-  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov_copy, transform = TRUE))[3, 3]), 0.1926955, tolerance = 0.000001)
-  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov_copy))[3, 3]), 0.1926955, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov_copy, transform = TRUE))[3, 3]), 0.1379893, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressWarnings(predict(cov, bills_cov_copy))[3, 3]), 0.1379893, tolerance = 0.000001)
 })
 
 
@@ -98,7 +111,7 @@ cov <- keyATM(docs = keyATM_docs,
               keywords = bills_keywords,
               model = "covariates",
               model_settings = list(covariates_data = bills_cov_modified, standardize = "none", 
-                                  covariates_formula = ~.
+                                  covariates_formula = ~., covariates_model = "PG"
                                   ),
               options = list(seed = 250, store_theta = TRUE, iterations = 5,
                              store_pi = 1, thinning = 5, verbose = FALSE)
@@ -111,7 +124,7 @@ test_that("Covariates settings: Standardize - none", {
 
   skip_on_os("linux") ; skip_on_cran()
   expect_error(predict(cov, bills_cov_modified))
-  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[3, 3]), 0.008454946, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[3, 3]), 3.90292e-07, tolerance = 0.000001)
 })
 
 
@@ -120,7 +133,7 @@ cov <- keyATM(docs = keyATM_docs,
               keywords = bills_keywords,
               model = "covariates",
               model_settings = list(covariates_data = bills_cov_modified, standardize = "non-factor", 
-                                  covariates_formula = ~.
+                                  covariates_formula = ~., covariates_model = "PG"
                                   ),
               options = list(seed = 250, store_theta = TRUE, iterations = 5,
                              store_pi = 1, thinning = 5, verbose = FALSE)
@@ -134,7 +147,7 @@ test_that("Covariates settings: Standardize - non-factor", {
 
   skip_on_os("linux") ; skip_on_cran()
   expect_error(predict(cov, bills_cov_modified))
-  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[2, 3]), 0.1179603, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[2, 3]), 0.124148, tolerance = 0.000001)
 })
 
 
@@ -144,7 +157,7 @@ cov <- keyATM(docs = keyATM_docs,
               keywords = bills_keywords,
               model = "covariates",
               model_settings = list(covariates_data = bills_cov_modified, standardize = "all", 
-                                  covariates_formula = ~.
+                                  covariates_formula = ~., covariates_model = "PG"
                                   ),
               options = list(seed = 250, store_theta = TRUE, iterations = 5,
                              store_pi = 1, thinning = 5, verbose = FALSE)
@@ -159,6 +172,6 @@ test_that("Covariates settings: Standardize - all", {
 
   skip_on_os("linux") ; skip_on_cran()
   expect_error(predict(cov, bills_cov_modified))
-  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[5, 2]), 0.086891, tolerance = 0.000001)
+  expect_equal(as.numeric(suppressMessages(predict(cov, bills_cov_modified, transform = TRUE))[5, 2]), 0.03749426, tolerance = 0.0001)
 })